requirements.txt
pyrogram~=1.2.9
pyromod
TgCrypto
SQLAlchemy~=1.4.21
psycopg2
instaloader~=4.8
python-dotenv
Flask